redo uninstall_sensor unit test
